Выполнил задание 1
