+++ geekbrains.level1/Lesson6/Program.cs
@@ -6,43 +6,16 @@ namespace Lesson6
 {
     class Program
     {
-        public static double F(double x)
-        {
-            return x * x - 50 * x + 10;
-        }
-        public static void SaveFunc(string fileName, double a, double b, double h)
-        {
-            FileStream fs = new FileStream(fileName, FileMode.Create, FileAccess.Write);
-            BinaryWriter bw = new BinaryWriter(fs);
-            double x = a;
-            while (x <= b)
-            {
-                bw.Write(F(x));
-                x += h;// x=x+h;
-            }
-            bw.Close();
-            fs.Close();
-        }
-        public static double Load(string fileName)
-        {
-            FileStream fs = new FileStream(fileName, FileMode.Open, FileAccess.Read);
-            BinaryReader bw = new BinaryReader(fs);
-            double min = double.MaxValue;
-            double d;
-            for (int i = 0; i < fs.Length / sizeof(double); i++)
-            {
-                // Считываем значение и переходим к следующему
-                d = bw.ReadDouble();
-                if (d < min) min = d;
-            }
-            bw.Close();
-            fs.Close();
-            return min;
-        }
+        
         static void Main(string[] args)
         {
-            SaveFunc("data.bin", -100, 100, 0.5);
-            Console.WriteLine(Load("data.bin"));
+            // 1. Продемонстрировать работу на функции с функцией a*x^2 и функцией a*sin(x).
+                        Func.SaveFunc(new FuncD(Func.F1),"data.bin", -100, 100, 0.5,3);
+            Console.WriteLine(Func.Load("data.bin"));
+            Func.SaveFunc(new FuncD(Func.F2), "data.bin", -100, 100, 0.5, 3);
+            Console.WriteLine(Func.Load("data.bin"));
+
+
             Console.ReadKey();
         }
     }
